some more functions

+++ communcore/libft/ft_strlen.c
@@ -1,8 +1,10 @@
 #include "libft.h"
 
+// this functions returns the lenght of the string str
+
 size_t	ft_strlen(const char *str)
 {
-	int	ctd;
+	size_t	ctd;
 
 	ctd = 0;
 	while (str[ctd])
